+++ example/android/app/src/main/java/com/posthog_flutter_example/MainActivity.java
@@ -1,0 +1,10 @@
+package com.posthog.posthog_flutter_example;
+
+
+import io.flutter.embedding.android.FlutterActivity;
+import io.flutter.embedding.engine.FlutterEngine;
+
+public class MainActivity extends FlutterActivity {
+  // You can keep this empty class or remove it. Plugins on the new embedding
+  // now automatically registers plugins.
+}