+++ ios/Classes/PosthogFlutterPlugin.h
@@ -1,0 +1,4 @@
+#import <Flutter/Flutter.h>
+
+@interface PosthogFlutterPlugin : NSObject <FlutterPlugin>
+@end