+++ lib/src/posthog_default_options.dart
@@ -1,0 +1,8 @@
+class PosthogDefaultOptions {
+  Map<String, dynamic>? options;
+
+  /// Singleton of [PosthogDefaultOptions].
+  static final PosthogDefaultOptions instance = PosthogDefaultOptions._();
+
+  PosthogDefaultOptions._();
+}